netwerk/test/mochitests/test_1425031.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1425031
-->
<head>
  <title>Cookies set in content processes update immediately.</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1425031">Mozilla Bug 1425031</a>
<p id="display"></p>
<div id="content" style="display: none">
<script type="application/javascript">

// Verify that cookie operations initiated by content processes do not cause
// asynchronous updates for those operations to be processed later.

SimpleTest.waitForExplicitFinish();

var gScript = SpecialPowers.loadChromeScript(SimpleTest.getTestFileURL('file_1331680.js'));
var testsNum = 0;
var cookieString = "cookie0=test";
let {COOKIE_ADDED, COOKIE_DELETED} = SpecialPowers.Ci.nsICookieNotification;

// Confirm the notify which represents the cookie is updating.
function confirmCookieOperation(op) {
  testsNum++;
  switch(testsNum) {
    case 1:
      is(op, COOKIE_ADDED, "Confirm the cookie operation is added.");
      is(document.cookie, cookieString, "Confirm the cookie string is unaffected by the addition");
      break;
    case 2:
      is(op, COOKIE_DELETED, "Confirm the cookie operation is deleted.");
      is(document.cookie, cookieString, "Confirm the cookie string is unaffected by the deletion");
      break;
    case 3:
      is(op, COOKIE_ADDED, "Confirm the cookie operation is added.");
      is(document.cookie, cookieString, "Confirm the cookie string is unaffected by the second addition.");
      document.cookie = "cookie0=; expires=Thu, 01-Jan-1970 00:00:01 GMT;";
      gScript.sendAsyncMessage('removeObserver');
      SpecialPowers.clearUserPref("network.cookie.sameSite.laxByDefault");
      SimpleTest.finish();
      break;
  }
}

function testSetCookie() {
  document.cookie = cookieString;
  is(document.cookie, cookieString, "Confirm cookie string.");
  document.cookie = "cookie0=; expires=Thu, 01-Jan-1970 00:00:01 GMT;";
  is(document.cookie, "", "Removed all cookies.");
  document.cookie = cookieString;
  is(document.cookie, cookieString, "Confirm cookie string.");
}

// Bug 1617611: Fix all the tests broken by "cookies SameSite=lax by default"
SpecialPowers.pushPrefEnv({
  set: [["network.cookie.sameSite.laxByDefault", false]],
}, () => {
  gScript.addMessageListener("cookieOperation", confirmCookieOperation);
  gScript.addMessageListener("createObserver:return", testSetCookie);
  gScript.sendAsyncMessage('createObserver');
});

</script>

</div>
<pre id="test">
</pre>
</body>
</html>
